c7

+++ lib/My_lib/Var.h
@@ -27,7 +27,7 @@ volatile unsigned long lastDebounceTime_4 = 0;
 const unsigned long debounceDelay = 300; // Thời gian chờ để xử lý nhiễu (milliseconds)
 uint8_t check_button_loop = 0;  // Tránh nhấn nhiều lần 1 nút
 
-uint8_t Set_time_day = 120 * 60 *1000;
+unsigned long Set_time_day = 120 * 60 * 1000;
 bool Bt_ready = false; // trạng thái bấm nút
 
 bool check_time = false;
